docs: repository/service/util 에 javadocs 작성
