+++ src/main/java/com/nhnacademy/bookstoreinjun/repository/BookRepository.java
@@ -13,24 +13,25 @@ import org.springframework.stereotype.Repository;
 
 public interface BookRepository extends JpaRepository<Book, Long> {
     /**
-     * @param pageable 웹에서 사용자가 전달할 pageable 인자.
-     * @return productState = 0인 (현재 판매 중인) 도서의 페이지.
+     * @param pageable 웹에서 사용자가 전달할 pageable 인자
+     * @return productState = 0인 (현재 판매 중인) 도서의 페이지
     */
     @Query("select b from Book b join b.product p where p.productState = 0")
     Page<Book> findBooksByProductState(Pageable pageable);
 
+
+    /**
+     * @param pageable 웹에서 사용자가 전달할 pageable 인자
+     * @param title 검색할 도서명
+     * @return 제목에 해당 도서명을 포함하면서 productState = 0인 (현재 판매 중인) 도서의 페이지
+     */
     @Query("select b from Book b join b.product p where p.productState = 0 and b.title like %:title%")
     Page<Book> findBooksByProductStateAndNameContaining(Pageable pageable, @Param("title") String title);
 
     /**
-     * @param isbn 도서의 isbn10 (각 도서별로 고유한 코드. 이걸로 세계 모든 도서를 구분할 수 있습니다.)
-     * @return 현재 데이터베이스에 해당 isbn13에 해당하는 도서가 존재하는 지 여부.
+     * 특정 isbn10 (각 도서별로 고유한 코드. 이걸로 세계 모든 도서를 구분할 수 있습니다.)에 해당하는 도서를 반환합니다.
+     * @param isbn 도서의 isbn10
+     * @return 현재 데이터베이스에 해당 isbn10에 해당하는 도서가 존재하는 지 여부
      */
     boolean existsByIsbn(String isbn);
-
-    /**
-     * @param product 도서가 가지는 product.
-     * @return 해당 product 에 대응되는 도서. 1대 1 관계이기 때문에 리스트가 아닌 단일 객체가 반환됩니다.
-     */
-    Book findByProduct(Product product);
 }
