+++ src/test/java/com/nhnacademy/bookstoreinjun/book/service/BookServiceTest.java
@@ -227,51 +227,6 @@ public class BookServiceTest {
         testProduct.setProductCategoryRelations(new ArrayList<>());
         testProduct.setProductTags(new ArrayList<>());
 
-        List<ProductCategoryRelation> testProductCategories = Arrays.asList(
-                ProductCategoryRelation.builder()
-                        .product(testProduct)
-                        .productCategory(
-                                ProductCategory.builder()
-                                        .categoryName("test category1")
-                                        .build()
-                        ).build(),
-                ProductCategoryRelation.builder()
-                        .product(testProduct)
-                        .productCategory(
-                                ProductCategory.builder()
-                                        .categoryName("test category2")
-                                        .build()
-                        ).build()
-        );
-
-        testProduct.setProductCategoryRelations(testProductCategories);
-
-        List<ProductTag> testProductTags = Arrays.asList(
-                ProductTag.builder()
-                        .product(testProduct)
-                        .tag(
-                                Tag.builder()
-                                        .tagName("test tag1")
-                                        .build()
-                        ).build(),
-                ProductTag.builder()
-                        .product(testProduct)
-                        .tag(
-                                Tag.builder()
-                                        .tagName("test tag2")
-                                        .build()
-                        ).build(),
-                ProductTag.builder()
-                        .product(testProduct)
-                        .tag(
-                                Tag.builder()
-                                        .tagName("test tag3")
-                                        .build()
-                        ).build()
-        );
-
-        testProduct.setProductTags(testProductTags);
-
         testBook.setProduct(testProduct);
 
         when(bookRepository.findById(1L)).thenReturn(Optional.of(testBook));
